Added Auth.js and Auth component to header
