+++ client/src/components/Header.js
@@ -1,6 +1,7 @@
 import React from "react";
 import "../styles/Header.css"
 import Nav from "./Nav"
+import Auth from "./Auth"
 
 function Header() {
   return (
@@ -8,6 +9,7 @@ function Header() {
     <header className="header">
       <img className="logo" width="50px" src="./assets/Image/logoPic.png" alt="logo"></img>
       <span className="wordmark pl-3">Mary Jane's Market</span>
+      <Auth />
       <div>
       <Nav />
       </div>
